Reorganized sitemap modules for proper testing
