+++ examples/next/getting-started/src/pages/_middleware.tsx
@@ -2,8 +2,8 @@ import 'faust.config';
 import { handleSitemapRequests } from '@faustjs/next';
 import { NextRequest, NextResponse } from 'next/server';
 
-export default function middleware(req: NextRequest) {
-  const isSitemapRequest = handleSitemapRequests(req, {
+export default async function middleware(req: NextRequest) {
+  const isSitemapRequest = await handleSitemapRequests(req, {
     sitemapIndexPath: '/sitemap_index.xml',
     sitemapPathsToIgnore: ['/author-*'],
     replaceUrls: true,
