+++ packages/next/src/server/sitemaps/handleSitemapRequests.ts
@@ -1,0 +1,178 @@
+import defaults from 'lodash/defaults.js';
+import isArray from 'lodash/isArray.js';
+import isBoolean from 'lodash/isBoolean.js';
+import isObject from 'lodash/isObject.js';
+import isString from 'lodash/isString.js';
+import isUndefined from 'lodash/isUndefined.js';
+import { NextRequest } from 'next/server.js';
+import {
+  createPagesSitemap,
+  createRootSitemapIndex,
+  handleSitemapPath,
+} from './createSitemaps.js';
+import { SitemapSchemaUrlElement } from './sitemapUtils.js';
+
+/**
+ * the "pages" config type
+ */
+export interface NextJSPage extends Omit<SitemapSchemaUrlElement, 'loc'> {
+  /**
+   * The relative URL of the Next.js page.
+   *
+   * @example /about
+   */
+  path: string;
+}
+
+export interface HandleSitemapRequestsConfig {
+  /**
+   * The pathname to the sitemap index file.
+   *
+   * @example /sitemap.xml, /wp-sitemap.xml, /sitemap_index.xml
+   */
+  sitemapIndexPath: string;
+  /**
+   * A list of pathnames to ignore when proxying sitemaps.
+   */
+  sitemapPathsToIgnore?: string[];
+  /**
+   * Next.js pages you want included in you sitemap. When provided, an index
+   * will be created specifically for these pages.
+   */
+  pages?: NextJSPage[];
+  /**
+   * Replace the WordPress site URL for your headless frontend site url in the
+   * sitemap url entries.
+   */
+  replaceUrls: boolean;
+}
+
+/**
+ * The pathname to the Next.js pages sitemap file. We may want to make this
+ * configurable in the future.
+ */
+export const FAUST_PAGES_PATHNAME = '/sitemap-faust-pages.xml';
+
+/**
+ * Validates the structure of the user defined config.
+ *
+ * @param {Partial<HandleSitemapRequestsConfig>} config The user provided config
+ */
+export function validateConfig(
+  config: Partial<HandleSitemapRequestsConfig>,
+): void {
+  if (isUndefined(config?.sitemapIndexPath)) {
+    throw new Error('sitemapIndexPath is required');
+  }
+
+  if (!isString(config?.sitemapIndexPath)) {
+    throw new Error('sitemapIndexPath must be a string');
+  }
+
+  if (!config?.sitemapIndexPath.startsWith('/')) {
+    throw new Error('sitemapIndexPath must start with a forward slash');
+  }
+
+  if (!isUndefined(config?.sitemapPathsToIgnore)) {
+    if (!isArray(config.sitemapPathsToIgnore)) {
+      throw new Error('sitemapPathsToIgnore must be an array');
+    }
+
+    (config?.sitemapPathsToIgnore).forEach((path) => {
+      if (!isString(path)) {
+        throw new Error('sitemapPathsToIgnore must be an array of strings');
+      }
+
+      if (!path.startsWith('/')) {
+        throw new Error(
+          'Each sitemapPathsToIgnore must start with a forward slash',
+        );
+      }
+
+      if (path.includes('*') && !path.endsWith('*')) {
+        throw new Error(
+          'sitemapPathsToIgnore with a wildcard must end with a wildcard',
+        );
+      }
+    });
+  }
+
+  // Validate pages structure and required values
+  if (!isUndefined(config?.pages)) {
+    if (!isArray(config.pages)) {
+      throw new Error('pages must be an array');
+    }
+
+    (config?.pages).forEach((page) => {
+      if (!isObject(page)) {
+        throw new Error('pages must be an array of objects');
+      }
+
+      if (isUndefined(page.path)) {
+        throw new Error('pages must have a path property');
+      }
+
+      if (!isString(page.path)) {
+        throw new Error('The pages path property must be a string');
+      }
+    });
+  }
+
+  // Validate replaceUrls is a boolean
+  if (!isUndefined(config?.replaceUrls) && !isBoolean(config?.replaceUrls)) {
+    throw new Error('replaceUrls must be a boolean');
+  }
+}
+
+/**
+ * Next.js middleware to proxy sitemap requests from the WordPress site.
+ *
+ * @param req The Next.js middleware request object
+ * @param config The user specified config object
+ * @returns {Response|undefined} A response object if the current request
+ * is for a sitemap that needs to be handled, undefined otherwise
+ */
+export async function handleSitemapRequests(
+  req: NextRequest,
+  config: Partial<HandleSitemapRequestsConfig>,
+): Promise<Response | undefined> {
+  // Validate config structure
+  validateConfig(config);
+
+  // Normalize config if some optional values are missing
+  const normalizedConfig: HandleSitemapRequestsConfig = defaults(
+    {},
+    config as HandleSitemapRequestsConfig,
+    {
+      replaceUrls: true,
+    },
+  );
+
+  const { pathname } = new URL(req.url);
+  const { sitemapIndexPath, pages } = normalizedConfig;
+
+  // Handle the root XML sitemap if specified in the config
+  if (pathname === sitemapIndexPath) {
+    return createRootSitemapIndex(req, normalizedConfig);
+  }
+
+  // Handle the sitemap for the specified Next.js pages if specified in the config
+  if (pathname === FAUST_PAGES_PATHNAME && pages?.length) {
+    return createPagesSitemap(req, normalizedConfig);
+  }
+
+  // Handle the sitemap index paths specified in the config
+  if (
+    pathname.includes('sitemap') &&
+    pathname.endsWith('.xml') &&
+    !normalizedConfig?.sitemapPathsToIgnore?.includes(pathname)
+  ) {
+    return handleSitemapPath(req, normalizedConfig);
+  }
+
+  /**
+   * If the above conditions are not met, return undefined and go on to the
+   * next middleware
+   */
+  return undefined;
+}
